Перенос интерфейсов, которые проксирует UI/Base, в пакующуюся библиотеку. https://online.sbis.ru/opendoc.html?guid=f0c35cf2-cd22-460c-a7f2-f2e2a8e8e29a
