+++ UI/_base/interface/IHTML.ts
@@ -1,4 +1,4 @@
-import {IControlOptions} from 'UICommon/interfaces';
+import {IControlOptions} from 'UICommon/Base';
 
 /**
  * Интерфейс для контролов, поддерживающих конфигурацию HTML-документа.
