+++ UICommon/Base.ts
@@ -1,1 +1,3 @@
 export { getProxyChildren } from './_base/ProxyChildren';
+
+export { IControlOptions, IControlChildren } from './_base/Control';
