+++ UIReact/UICore/_base/Control.ts
@@ -19,7 +19,8 @@ import { OptionsResolver } from 'UICommon/Executor';
 
 import { WasabyEvents, callNotify } from 'UICore/Events';
 import { IWasabyEventSystem } from 'UICommon/Events';
-import { TIState, IControlOptions } from 'UICommon/interfaces';
+import { TIState } from 'UICommon/interfaces';
+import { IControlOptions } from 'UICommon/Base';
 
 /**
  * Базовый контрол, наследник React.Component с поддержкой совместимости с Wasaby
